% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.mcemGLMM.R
\name{coef.mcemGLMM}
\alias{coef}
\alias{coef.mcemGLMM}
\title{Fixed effect coefficients extraction for mcemGLMM objects}
\usage{
\method{coef}{mcemGLMM}(object, ...)
}
\arguments{
\item{object}{a model fitted with the mcemGLMM function.}

\item{...}{additional arguments.}
}
\value{
A vector with the fixed effect coefficients.
}
\description{
Extract the fixed effect coeffcients from a model fitted with
\code{mcemGLMM}.
}
\author{
Felipe Acosta Archila <acosta@umn.edu>
}
\keyword{glmm}

