% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.mcemGLMM.R
\name{fitted.mcemGLMM}
\alias{fitted}
\alias{fitted.mcemGLMM}
\alias{fitted.values}
\title{Extract fitted values for an mcemGLMM object}
\usage{
\method{fitted}{mcemGLMM}(object, ...)
}
\arguments{
\item{object}{a model fitted with the mcemGLMM function.}

\item{...}{additional arguments.}
}
\value{
A vector of fitted values.
}
\description{
This function returns the fitted values of a model fitted with \code{mcemGLMM}.
}

