% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.mcemGLMM.R
\name{ranef.mcemGLMM}
\alias{ranef}
\alias{ranef.mcemGLMM}
\title{Random effect prediction for mcemGLMM objects}
\usage{
ranef.mcemGLMM(object, ...)
}
\arguments{
\item{object}{a model fitted with the mcemGLMM function.}

\item{...}{additional arguments.}
}
\value{
A vector with the random effect predictions.
}
\description{
Extract the random effect predictions from a model fitted with
\code{mcemGLMM}.
}
\author{
Felipe Acosta Archila <acosta@umn.edu>
}
\keyword{glmm}

