% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.mcemGLMM.R
\name{vcov.mcemGLMM}
\alias{vcov}
\alias{vcov.mcemGLMM}
\title{Fixed effect covariance matrix for an mcemGLMM object}
\usage{
\method{vcov}{mcemGLMM}(object, ...)
}
\arguments{
\item{object}{a model fitted with the mcemGLMM function.}

\item{...}{additional arguments.}
}
\value{
A matrix corresponding to the covariance matrix of the fixed
effects.
}
\description{
Extract the fixed effect's covariance matrix from a model fitted with
\code{mcemGLMM}.
}
\author{
Felipe Acosta Archila <acosta@umn.edu>
}
\keyword{glmm}

