\name{coef.mcemGLMM}
\alias{coef}
\alias{coef.mcemGLMM}
\title{Anova method for mcemGLMM objects}
\description{
  Extract the fixed effect coeffcients from a model fitted with \code{mcemGLMM}.
}
\usage{
\method{coef}{mcemGLMM}(object, ...)
}
\arguments{
  \item{object}{a model fitted with the mcemGLMM function.}
  \item{...}{additional arguments.}
}
\value{
  A vector with the fixed effect coefficients.
}
\author{
Felipe Acosta Archila <acosta@umn.edu>
}
\keyword{glmm}
