\name{ranef.mcemGLMM}
\alias{ranef}
\alias{ranef.mcemGLMM}
\title{Anova method for mcemGLMM objects}
\description{
  Extract the random effect predictions from a model fitted with \code{mcemGLMM}.
}
\usage{
\method{ranef}{mcemGLMM}(object, ...)
}
\arguments{
  \item{object}{a model fitted with the mcemGLMM function.}
  \item{...}{additional arguments.}
}
\value{
  A vector with the random effect predictions.
}
\author{
Felipe Acosta Archila <acosta@umn.edu>
}
\keyword{glmm}

