\name{residuals.mcemGLMM}
\alias{residuals}
\alias{residuals.mcemGLMM}
\title{Residual extraction method for mcemGLMM objects}
\description{
  This functions returns the residuals of a model fitted with \code{mcemGLMM}.
}
\usage{
\method{residuals}{mcemGLMM}(object, ...)
}
\arguments{
  \item{object}{a model fitted with the mcemGLMM function.}
  \item{...}{additional arguments.}
}
\value{
  A vector with the residuals of the model.
}
\details{
This section will state how is this obtained.
}
\keyword{glmm, residuals}
