\name{salamander}
\alias{salamander}
\docType{data}
\title{
Salamander Data
}
\description{
Salamander mating experiment data from McCullagh and Nelder.
}
\usage{data("salamander")}
\format{
  A data frame with 60 observations on the following 4 variables.
  \describe{
    \item{\code{Cross}}{the type of salamanders involved in the crossing.}
    \item{\code{Female}}{the female salamander's ID.}
    \item{\code{Male}}{the male salamander's ID.}
    \item{\code{Mate}}{wheter the crossing was succesful or not.}
  }
}

\references{
McCullagh, P and Nelder J. A. (1989) Generalized Linear Models, Second Edition. Chapman \& Hall, 1989
}
\keyword{datasets}
